+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - Removed extra `.graphiql-container` wrapper element
 - Added explorer open state saving to localStorage
 - Added a resizer for explorer and tweaked resizer's appearance
+- Changed main exposing function to `createGraphiqlApp(endpoint, discovery, options)`; `options` is supported `rootEl` and `title` options for now.
 
 ## 1.2.0 (10-03-2021)